# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Doctor {
  id: number;
  name: string;
  specialty: string;
  initials: string;
  accent: string;
  nextAvailable: string;
  availableSlots?: string[];
}

export type AppointmentStatus = typeof AppointmentStatus[keyof typeof AppointmentStatus];


export const AppointmentStatus = {
  upcoming: 'upcoming',
  completed: 'completed',
  cancelled: 'cancelled',
} as const;

export interface Appointment {
  id: number;
  doctorId: number;
  doctorName: string;
  specialty: string;
  date: string;
  time: string;
  reason: string;
  status: AppointmentStatus;
}

export interface AppointmentInput {
  /** @minimum 1 */
  doctorId: number;
  /** @minLength 1 */
  date: string;
  /** @minLength 1 */
  time: string;
  /** @minLength 3 */
  reason: string;
}

export type AppointmentUpdateStatus = typeof AppointmentUpdateStatus[keyof typeof AppointmentUpdateStatus];


export const AppointmentUpdateStatus = {
  upcoming: 'upcoming',
  completed: 'completed',
  cancelled: 'cancelled',
} as const;

export interface AppointmentUpdate {
  status?: AppointmentUpdateStatus;
  date?: string;
  time?: string;
  reason?: string;
}

export type MedicalRecordRecordType = typeof MedicalRecordRecordType[keyof typeof MedicalRecordRecordType];


export const MedicalRecordRecordType = {
  visit: 'visit',
  lab: 'lab',
  prescription: 'prescription',
  note: 'note',
} as const;

export type MedicalRecordStatus = typeof MedicalRecordStatus[keyof typeof MedicalRecordStatus];


export const MedicalRecordStatus = {
  active: 'active',
  archived: 'archived',
} as const;

export interface MedicalRecord {
  id: number;
  recordType: MedicalRecordRecordType;
  title: string;
  provider: string;
  date: string;
  summary: string;
  status: MedicalRecordStatus;
}

export type RecordInputRecordType = typeof RecordInputRecordType[keyof typeof RecordInputRecordType];


export const RecordInputRecordType = {
  visit: 'visit',
  lab: 'lab',
  prescription: 'prescription',
  note: 'note',
} as const;

export interface RecordInput {
  recordType: RecordInputRecordType;
  /** @minLength 1 */
  title: string;
  /** @minLength 1 */
  provider: string;
  /** @minLength 1 */
  date: string;
  /** @minLength 1 */
  summary: string;
}

export type RecordUpdateStatus = typeof RecordUpdateStatus[keyof typeof RecordUpdateStatus];


export const RecordUpdateStatus = {
  active: 'active',
  archived: 'archived',
} as const;

export interface RecordUpdate {
  title?: string;
  provider?: string;
  date?: string;
  summary?: string;
  status?: RecordUpdateStatus;
}

export type PredictionUrgency = typeof PredictionUrgency[keyof typeof PredictionUrgency];


export const PredictionUrgency = {
  routine: 'routine',
  attention: 'attention',
  urgent: 'urgent',
} as const;

export interface Prediction {
  id: number;
  createdAt: string;
  symptoms: string[];
  condition: string;
  /**
     * @minimum 0
     * @maximum 100
     */
  confidence: number;
  associatedSymptoms: string[];
  recommendation: string;
  urgency: PredictionUrgency;
}

export interface PredictionInput {
  /** @minItems 1 */
  symptoms: string[];
  otherSymptoms?: string;
}

export interface Dashboard {
  patientName: string;
  totalAppointments: number;
  recordsCount: number;
  upcomingAppointment: Appointment | null;
  recentPrediction: Prediction | null;
  wellnessScore: number;
  profileCompletion: number;
}

export type AdminOverviewAppointmentsByStatusItem = {
  label: string;
  value: number;
};

export interface AdminOverview {
  patientCount: number;
  appointmentCount: number;
  pendingAppointments: number;
  recordsAddedThisWeek: number;
  appointmentsByStatus: AdminOverviewAppointmentsByStatusItem[];
}