# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: dashboard
    description: Patient and clinician dashboard summaries
  - name: doctors
    description: Doctor directory and availability
  - name: appointments
    description: Appointment booking and management
  - name: records
    description: Digital patient records
  - name: predictions
    description: Symptom-based guidance and prediction history
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /dashboard:
    get:
      operationId: getDashboard
      tags: [dashboard]
      summary: Get the patient dashboard
      responses:
        "200":
          description: Dashboard summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Dashboard"
  /admin/overview:
    get:
      operationId: getAdminOverview
      tags: [dashboard]
      summary: Get the clinician dashboard overview
      responses:
        "200":
          description: Clinician dashboard summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminOverview"
  /doctors:
    get:
      operationId: listDoctors
      tags: [doctors]
      summary: List doctors and their availability
      responses:
        "200":
          description: Doctor directory
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Doctor"
  /appointments:
    get:
      operationId: listAppointments
      tags: [appointments]
      summary: List appointments for the current patient
      responses:
        "200":
          description: Appointment list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Appointment"
    post:
      operationId: createAppointment
      tags: [appointments]
      summary: Book an appointment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AppointmentInput"
      responses:
        "201":
          description: Appointment created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appointment"
  /appointments/{id}:
    patch:
      operationId: updateAppointment
      tags: [appointments]
      summary: Update or cancel an appointment
      parameters:
        - $ref: "#/components/parameters/Id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AppointmentUpdate"
      responses:
        "200":
          description: Updated appointment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appointment"
        "404":
          description: Appointment not found
  /records:
    get:
      operationId: listRecords
      tags: [records]
      summary: List digital patient records
      responses:
        "200":
          description: Medical records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MedicalRecord"
    post:
      operationId: createRecord
      tags: [records]
      summary: Add a medical record
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RecordInput"
      responses:
        "201":
          description: Record created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MedicalRecord"
  /records/{id}:
    patch:
      operationId: updateRecord
      tags: [records]
      summary: Update a medical record
      parameters:
        - $ref: "#/components/parameters/Id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RecordUpdate"
      responses:
        "200":
          description: Updated record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MedicalRecord"
        "404":
          description: Record not found
    delete:
      operationId: deleteRecord
      tags: [records]
      summary: Delete a medical record
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "204":
          description: Record deleted
        "404":
          description: Record not found
  /predictions:
    get:
      operationId: listPredictions
      tags: [predictions]
      summary: List symptom prediction history
      responses:
        "200":
          description: Prediction history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Prediction"
    post:
      operationId: createPrediction
      tags: [predictions]
      summary: Generate symptom-based guidance
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PredictionInput"
      responses:
        "201":
          description: Prediction guidance
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Prediction"
components:
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema:
        type: integer
        minimum: 1
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    Doctor:
      type: object
      required: [id, name, specialty, initials, accent, nextAvailable]
      properties:
        id: { type: integer }
        name: { type: string }
        specialty: { type: string }
        initials: { type: string }
        accent: { type: string }
        nextAvailable: { type: string }
        availableSlots:
          type: array
          items: { type: string }
    Appointment:
      type: object
      required: [id, doctorId, doctorName, specialty, date, time, reason, status]
      properties:
        id: { type: integer }
        doctorId: { type: integer }
        doctorName: { type: string }
        specialty: { type: string }
        date: { type: string }
        time: { type: string }
        reason: { type: string }
        status: { type: string, enum: [upcoming, completed, cancelled] }
    AppointmentInput:
      type: object
      required: [doctorId, date, time, reason]
      properties:
        doctorId: { type: integer, minimum: 1 }
        date: { type: string, minLength: 1 }
        time: { type: string, minLength: 1 }
        reason: { type: string, minLength: 3 }
    AppointmentUpdate:
      type: object
      properties:
        status: { type: string, enum: [upcoming, completed, cancelled] }
        date: { type: string }
        time: { type: string }
        reason: { type: string }
    MedicalRecord:
      type: object
      required: [id, recordType, title, provider, date, summary, status]
      properties:
        id: { type: integer }
        recordType: { type: string, enum: [visit, lab, prescription, note] }
        title: { type: string }
        provider: { type: string }
        date: { type: string }
        summary: { type: string }
        status: { type: string, enum: [active, archived] }
    RecordInput:
      type: object
      required: [recordType, title, provider, date, summary]
      properties:
        recordType: { type: string, enum: [visit, lab, prescription, note] }
        title: { type: string, minLength: 1 }
        provider: { type: string, minLength: 1 }
        date: { type: string, minLength: 1 }
        summary: { type: string, minLength: 1 }
    RecordUpdate:
      type: object
      properties:
        title: { type: string }
        provider: { type: string }
        date: { type: string }
        summary: { type: string }
        status: { type: string, enum: [active, archived] }
    Prediction:
      type: object
      required: [id, createdAt, symptoms, condition, confidence, associatedSymptoms, recommendation, urgency]
      properties:
        id: { type: integer }
        createdAt: { type: string }
        symptoms:
          type: array
          items: { type: string }
        condition: { type: string }
        confidence: { type: integer, minimum: 0, maximum: 100 }
        associatedSymptoms:
          type: array
          items: { type: string }
        recommendation: { type: string }
        urgency: { type: string, enum: [routine, attention, urgent] }
    PredictionInput:
      type: object
      required: [symptoms]
      properties:
        symptoms:
          type: array
          minItems: 1
          items: { type: string }
        otherSymptoms: { type: string }
    Dashboard:
      type: object
      required: [patientName, totalAppointments, recordsCount, upcomingAppointment, recentPrediction, wellnessScore, profileCompletion]
      properties:
        patientName: { type: string }
        totalAppointments: { type: integer }
        recordsCount: { type: integer }
        upcomingAppointment:
          anyOf:
            - $ref: "#/components/schemas/Appointment"
            - type: "null"
        recentPrediction:
          anyOf:
            - $ref: "#/components/schemas/Prediction"
            - type: "null"
        wellnessScore: { type: integer }
        profileCompletion: { type: integer }
    AdminOverview:
      type: object
      required: [patientCount, appointmentCount, pendingAppointments, recordsAddedThisWeek, appointmentsByStatus]
      properties:
        patientCount: { type: integer }
        appointmentCount: { type: integer }
        pendingAppointments: { type: integer }
        recordsAddedThisWeek: { type: integer }
        appointmentsByStatus:
          type: array
          items:
            type: object
            required: [label, value]
            properties:
              label: { type: string }
              value: { type: integer }