# lib/api-zod/src/generated/types/adminOverview.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AdminOverviewAppointmentsByStatusItem } from './adminOverviewAppointmentsByStatusItem';

export interface AdminOverview {
  patientCount: number;
  appointmentCount: number;
  pendingAppointments: number;
  recordsAddedThisWeek: number;
  appointmentsByStatus: AdminOverviewAppointmentsByStatusItem[];
}