# lib/api-zod/src/generated/types/appointmentUpdate.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AppointmentUpdateStatus } from './appointmentUpdateStatus';

export interface AppointmentUpdate {
  status?: AppointmentUpdateStatus;
  date?: string;
  time?: string;
  reason?: string;
}