# lib/api-zod/src/generated/types/dashboard.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Appointment } from './appointment';
import type { Prediction } from './prediction';

export interface Dashboard {
  patientName: string;
  totalAppointments: number;
  recordsCount: number;
  upcomingAppointment: Appointment | null;
  recentPrediction: Prediction | null;
  wellnessScore: number;
  profileCompletion: number;
}