# lib/api-zod/src/generated/types/medicalRecord.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { MedicalRecordRecordType } from './medicalRecordRecordType';
import type { MedicalRecordStatus } from './medicalRecordStatus';

export interface MedicalRecord {
  id: number;
  recordType: MedicalRecordRecordType;
  title: string;
  provider: string;
  date: string;
  summary: string;
  status: MedicalRecordStatus;
}