# lib/api-zod/src/generated/types/prediction.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { PredictionUrgency } from './predictionUrgency';

export interface Prediction {
  id: number;
  createdAt: string;
  symptoms: string[];
  condition: string;
  /**
     * @minimum 0
     * @maximum 100
     */
  confidence: number;
  associatedSymptoms: string[];
  recommendation: string;
  urgency: PredictionUrgency;
}