# lib/api-zod/src/generated/types/predictionInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface PredictionInput {
  /** @minItems 1 */
  symptoms: string[];
  otherSymptoms?: string;
}