# lib/api-zod/src/generated/types/recordInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { RecordInputRecordType } from './recordInputRecordType';

export interface RecordInput {
  recordType: RecordInputRecordType;
  /** @minLength 1 */
  title: string;
  /** @minLength 1 */
  provider: string;
  /** @minLength 1 */
  date: string;
  /** @minLength 1 */
  summary: string;
}